requirements.txt
Django==2.0.5
django-ckeditor==5.4.0
django-jet==1.0.7
django-js-asset==1.1.0
Pillow==5.1.0
psycopg2==2.7.4
pytz==2018.4
